requirements_external.txt
# Image
pillow

# Qr code
segno